# Conflicts:
#	src/Model/Shapes/Arrow.java
#	src/Model/Shapes/Circle.java
#	src/Model/Shapes/CurvedLine.java
#	src/Model/Shapes/Link.java
#	src/Model/Shapes/Node.java
#	src/Model/Shapes/Rectangle.java
#	src/Model/Shapes/Shape.java

+++ src/Model/Shapes/Line.java
@@ -1,7 +1,7 @@
 package Model.Shapes;
 
 public class Line extends Link {
-    public Line(int xOrigin, int yOrigin, int xDestination, int yDestination) {
+    public Line(float xOrigin, float yOrigin, float xDestination, float yDestination) {
         super(xOrigin, yOrigin, xDestination, yDestination);
     }
